implementing new register interface as prototype is
