+++ src/Pages/Register/Register.css
@@ -26,15 +26,16 @@
 
 .reg {
 
-    background-color: white;
+    background-color:#2B2B2B;
     width: 33%;
     height: 100%;
     display: flex;
     flex-direction: row;
     justify-content: center;
+    align-items: center;
 }
 
-.pull-everybody {
+.register-pull-everybody {
     width: 100%;
     display: flex;
     flex-direction: column;
@@ -44,8 +45,8 @@
 
 }
 
-.pull-everybody label {
-    color: #2B2B2B;
+.register-pull-everybody label {
+    color: #4F4F4F;
     font-family: Roboto;
     margin: 1.5% 0;
 }
