+++ src/Pages/Register/RegisterStep1.css
@@ -1,0 +1,63 @@
+@media screen and (max-width: 528px){
+    .container{
+        width: 100%;
+      
+    }
+
+    .fitting{
+        display: none;
+       
+    }
+
+    .reg_step1{
+        min-width: 100%;
+    }
+}
+.container {
+    max-width: 100%;
+    background-color: #2B2B2B;
+    display: flex;
+    max-height: fit-content;
+}
+
+.fitting {
+    width: 67%;
+    height: 100%;
+}
+
+.reg_step1 {
+    background-color:#2B2B2B;
+    width: 33%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.step1-content{
+    width: 100%;
+    padding: 10%;
+}
+
+.email-content{
+   display: flex;
+   flex-direction: column;
+}
+.fit-email-content{
+    display: flex;
+   flex-direction: row;
+   align-items: center;
+   justify-content:space-between;
+   color: white;
+   text-transform: uppercase;
+   font-size: 12px;
+   margin-top: 2%;
+   
+}
+
+.trace{
+    border: 0.5px solid #FFFFFF;
+    width: 45%;
+}
+
