+++ src/Shared/img/LogoSVG.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { SVGProps } from "react";
+
+const SVGLogo = ({ ...props }: SVGProps<SVGSVGElement>) => {
+    return (
+        <svg width="388" height="397" viewBox="0 0 388 397" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+            <path d="M193.5 0L361.076 208.5H25.9241L193.5 0Z" fill="#54D0F7" />
+            <path d="M194 6L306.583 183H81.4167L194 6Z" fill="#B5F2FF" />
+            <path d="M193.5 290.997L25.9341 207.747L361.086 207.787L193.5 290.997Z" fill="#54D0F6" />
+            <path d="M194 273L306.583 183H81.4167L194 273Z" fill="#B5F2FF" />
+            <path d="M152 313.7C157.4 313.7 161.967 314.2 165.7 315.2C169.433 316.2 172.4 318.133 174.6 321C176.867 323.8 178 327.833 178 333.1C178 335.433 177.633 337.833 176.9 340.3C176.233 342.7 175.367 344.8 174.3 346.6C173.3 348.4 172.4 349.567 171.6 350.1C170.667 350.767 168.6 351.3 165.4 351.7C162.2 352.033 158.433 352.2 154.1 352.2C152.567 352.2 150.2 352.133 147 352V371H129V314C136 313.8 143.667 313.7 152 313.7ZM159.5 339.4C160.167 338.533 160.5 336.8 160.5 334.2C160.5 330.8 159.767 328.4 158.3 327C156.9 325.6 154.433 324.9 150.9 324.9C149.7 324.9 148.4 325 147 325.2V340.4C147.667 340.467 148.967 340.5 150.9 340.5C156.033 340.5 158.9 340.133 159.5 339.4ZM200.81 314V350.5C200.81 355.567 200.343 359.667 199.41 362.8C198.477 365.933 196.81 368.3 194.41 369.9C192.077 371.5 188.777 372.3 184.51 372.3C179.977 372.3 175.21 371.867 170.21 371L171.71 359.5C173.577 359.567 175.71 359.533 178.11 359.4C180.51 359.267 181.877 359 182.21 358.6C182.61 357.933 182.81 356.333 182.81 353.8C182.81 351.267 182.81 349.833 182.81 349.5V314H200.81ZM219.941 341.8L206.541 314H224.941L234.141 336.4L243.641 314H261.541L248.341 341.8L262.441 371.1H244.041L234.141 346.9L223.941 371.1H205.441L219.941 341.8Z" fill="#54D0F6" />
+        </svg>
+    )
+}
+
+export default SVGLogo;
