+++ src/react-app-env.d.ts
@@ -1,0 +1,5 @@
+declare module '*.mp4' {
+    const src: string;
+    export default src;
+  }
+
